Vim: Add vim-move plugin
With mappings using ESC instead of Alt, because Alt doesn’t work in the
term, and we configured iterm2 to send +Esc escape sequence for
alt/option.

But because we use <ESC> in visual mode mappings, Vim won’t leave
visual mode instantly and rather wait for new touches to be typed
(especially painful when <ESC>k will move selection up instead of
exiting visual and moving cursor up one line). This will be addressed
in a later commit with the fast-keycodes hack.