Vim: Fix <Esc><BS> shortcut in CtrlP mappings
We applied the fast-keycodes hack (see
http://vim.wikia.com/wiki/Mapping_fast_keycodes_in_terminal_Vim)
earlier to `<Esc><BS>` in console mode, so now vim won’t accept
`<Esc><BS>` in a mapping, but only the vim keycode it was assigned to
(`<F16>`)

This should be kept in sync with the vim keycode used for `<Esc><BS>`
(currently `set <F16>=^[^?`) If that keycode changes, say to `<F17>`,
the change should be reflected here as well. If the mapping is removed,
then this commit can be reverted.